Address review: invert the refusal guard and share the CSM skip helper
The guard now reads in the same direction as its name and as the check in
uploadMedia(). The client-side skip gate the new spec needed already existed
as a method on the CSM spec's utils class; both now call one exported helper
so the two gates cannot drift.
